pyproject.toml
[project]
name = "Speech2Image"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.41.1",
    "replicate==1.0.4",
    "google-cloud-speech==2.29.0",
    "python-dotenv==1.0.1",
    "ruff>=0.8.5",
]
